Add OpenAI integration for product parsing and ingestion, including image-to-items processing, unknown product resolution, and nutrition data normalization functionalities; update schemas, repositories, services, and dependencies

+++ src/calorie/orm.py
@@ -73,7 +73,7 @@ class DayProduct(Base):
         primary_key=True,
     )
 
-    weight: Mapped[Decimal] = mapped_column(nullable=False)
+    weight: Mapped[int] = mapped_column(nullable=False)
 
     day: Mapped["Day"] = relationship("Day", back_populates="day_products")
     product: Mapped["Product"] = relationship("Product", back_populates="day_products")
